Extract call for requested resources to service
