+++ cloudapp/src/app/print-slip-report/print-slip-report.component.ts
@@ -5,7 +5,8 @@ import { AlertService } from '@exlibris/exl-cloudapp-angular-lib'
 import { Subscription } from 'rxjs'
 import { COLUMNS_DEFINITIONS } from '../column-definitions'
 import { ColumnOption } from '../column-options'
-import { PrintSlipReportService, RequestedResource } from './print-slip-report.service'
+import { RequestedResource } from '../requested-resources-service/requested-resources-service'
+import { PrintSlipReportService } from './print-slip-report.service'
